Работа над ошибками

+++ src/main/java/hh/homework/taskManager/controllers/TaskController.java
@@ -5,10 +5,15 @@ import hh.homework.taskManager.repositories.TaskRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ModelAttribute;
 
 import java.time.LocalDate;
-import java.util.UUID;
+import java.util.Optional;
 
 @Controller
 public class TaskController {
@@ -17,39 +22,49 @@ public class TaskController {
     public TaskController(TaskRepository taskRepository){
         this.taskRepository = taskRepository;
     }
-    @GetMapping("/home")
+    @GetMapping("/tasks")
     public String findAll(Model model){
         var tasks = taskRepository.findAll();
         model.addAttribute("tasks", tasks);
-        return "home/home.html";
+        return "tasks/tasks";
     }
-    @PostMapping("/home")
+    @PostMapping("/tasks")
     public String addTask(Model model, @RequestParam String taskName, @RequestParam(required = false) LocalDate deadline){
-        Task task = new Task(taskName, LocalDate.now(), deadline != null ? deadline : null, "Started");
+        Task task = new Task(taskName, deadline);
         taskRepository.save(task);
         var tasks = taskRepository.findAll();
         model.addAttribute("tasks", tasks);
-        return "home/home.html";
+        return "tasks/tasks";
     }
-    @DeleteMapping("/home")
+    @DeleteMapping("/tasks")
     public String deleteTask(Model model, @RequestParam("id") String stringId){
-        UUID id = UUID.fromString(stringId);
-        taskRepository.deleteById(id);
+        Integer id = Integer.parseInt(stringId);
+        if(taskRepository.existsById(id))
+            taskRepository.deleteById(id);
         var tasks = taskRepository.findAll();
         model.addAttribute("tasks", tasks);
-        return "home/home.html";
+        return "tasks/tasks";
     }
     @GetMapping("/edit")
     public String editTask(Model model, @RequestParam(name = "id") String stringId){
-        UUID id = UUID.fromString(stringId);
-        model.addAttribute("task", taskRepository.getReferenceById(id));
-        return "home/edit.html";
+        Integer id = Integer.parseInt(stringId);
+        if (taskRepository.existsById(id)) {
+            Optional<Task> taskToBeFoundOpt = taskRepository.findById(id);
+            Task taskToBeFound = taskToBeFoundOpt.get();
+            model.addAttribute("task", taskToBeFound);
+            return "tasks/edit";
+        } else
+            return "tasks/error";
     }
     @PatchMapping("/edit")
     public String updateTask(@ModelAttribute("task") Task task, Model model){
-        taskRepository.save(task);
-        var tasks = taskRepository.findAll();
-        model.addAttribute("tasks", tasks);
-        return "redirect:home";
+        if(taskRepository.existsById(task.getId())) {
+            taskRepository.save(task);
+            var tasks = taskRepository.findAll();
+            //tasks.sort(Comparator.comparing(Task::getTaskName));
+            model.addAttribute("tasks", tasks);
+            return "redirect:tasks";
+        } else
+            return "tasks/error";
     }
 }
